Tests/LibWeb/Text/expected/DOM/Document-createCDATASection.txt

<xml><![CDATA[Some <CDATA> data & then some]]></xml>
Exception: NotSupportedError
Exception: InvalidCharacterError
